Log events to file

+++ ground_station/Utilities/GraphUtils.h
@@ -2,11 +2,11 @@
 #define GS_MATTERHORN_GRAPHUTILS_H
 
 QCPGraphData speedFromReading(TelemetryReading reading) {
-    return {reading.timestamp, reading.speed.value};
+    return {static_cast<double>(reading.timestamp), reading.speed.value};
 }
 
 QCPGraphData accelerationFromReading(TelemetryReading reading) {
-    return {reading.timestamp, reading.acceleration.value};
+    return {static_cast<double>(reading.timestamp), reading.acceleration.value};
 }
 
 #endif //GS_MATTERHORN_GRAPHUTILS_H
